// src/app/shared/models/students/student-profile-update-request.ts
export class UpdateInformationResponce{
    id?: number;
    requestId?: string;
    studentId?: string;
    propertyName?: string;
    propertyValue?: string;
    documentType?: string;
    isDocumentUploaded?: string;
    status?: string;
    comments?: string;
    isApproved?: boolean;
    reviewerComments?: string;
}